+++ SQL/mysql.initial.sql
@@ -260,4 +260,4 @@ CREATE TABLE `system` (
 
 /*!40014 SET FOREIGN_KEY_CHECKS=1 */;
 
-INSERT INTO `system` (`name`, `value`) VALUES ('roundcube-version', '2021081000');
+INSERT INTO `system` (`name`, `value`) VALUES ('roundcube-version', '2021100300');